+++ CHANGES.md
@@ -5,6 +5,8 @@ Changes in 9.2.1
 ----------------
 - Remove use of db_seq_nextval(), use auto increment, program wide
 - SQL set default nextval (auto increment) for RosarioSIS version < 5.0 on install & old add-ons in Update.fnc.php
+- SQL no more cast MARKING_PERIOD_ID column as text/varchar in rosariosis.sql & InputFinalGrades.php
+- PLpgSQL compact & consistent function declaration in rosariosis.sql
 
 Changes in 9.2
 --------------